Chat Controller + Websocket hinzugefügt

+++ src/main/java/com/caro/IdeaExplorer/model/AuthorDto.java
@@ -8,10 +8,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class AuthorDto {
-    private Long id;
+    private Integer id;
     private String name;
 
-    public AuthorDto(Long id, String name) {
+    public AuthorDto(Integer id, String name) {
         this.id = id;
         this.name = name;
     }
